+++ tutu-backend/src/main/scala/fi/oph/tutu/backend/security/AuthenticationEventListener.scala
@@ -10,10 +10,9 @@ import org.springframework.security.authentication.event.AuthenticationSuccessEv
 import org.springframework.stereotype.Component
 import org.springframework.web.context.request.{RequestContextHolder, ServletRequestAttributes}
 
-
 @Component class AuthenticationEventListener(auditLog: AuditLog) {
   @EventListener def onAuthenticationSuccess(event: AuthenticationSuccessEvent): Unit = {
-    val target = new Target.Builder().setField("userOid", event.getAuthentication.getName).build()
+    val target  = new Target.Builder().setField("userOid", event.getAuthentication.getName).build()
     val request = getCurrentHttpRequest
     audit.log(getUser(request), Login, target, Changes.EMPTY)
     val username = event.getAuthentication.getName
